feate(mods): Create add ons (#74)
* feat(mods): create additions

crafts and additions. The bridge between kinetic energy and electricity. Adds some cool items we can use as well.

* feat(mods): create encased

allows all casings on shafts and cogs and machines and gearboxes.

* feat(mods): create sifter.

Sifting recipes should be useful.

* feat(mods): mechanical extruder

Adds a create version of thermals igneous extruder.

* feat(mods): create chunk loaders.

loads chunks with the power of create.